requirements.txt
# Dependencies for CARDIA RAG system
pandas>=1.5.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
tqdm>=4.64.0
google-generativeai>=0.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
